+++ src/games/ttw/falloutttwbsainvalidation.cpp
@@ -10,7 +10,7 @@ QString FalloutTTWBSAInvalidation::invalidationBSAName() const
   return "Fallout - Invalidation.bsa";
 }
 
-unsigned long FalloutTTWBSAInvalidation::bsaVersion() const
+uint32_t FalloutTTWBSAInvalidation::bsaVersion() const
 {
   return 0x68;
 }
